% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{apply_respfun_outcome}
\alias{apply_respfun_outcome}
\title{Apply link functions based on outcome information}
\usage{
apply_respfun_outcome(x, outcome, links)
}
\arguments{
\item{x}{Vector of additive predictors.}

\item{outcome}{Factor vector containing information on the outcome of
the corresponding element of vector x.}

\item{links}{Vector containing the names of the respective links for the mu
outcomes.}
}
\value{
Vector of lenght x where different response functions have been
  applied.
}
\description{
This is an internal function for the extreme case that a vector is plugged
into a response function depending on outcome information.
}
