% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{compress_outcomes}
\alias{compress_outcomes}
\title{Compress the outcome list of predictions into single vectors}
\usage{
compress_outcomes(pred_list, mus, sigmas, outcome)
}
\arguments{
\item{pred_list}{List of predictions for each outcome.}

\item{mus}{Character vector with names of included mu models.}

\item{sigmas}{Character vector with names of included sigma models.}

\item{outcome}{Factor vector containing the information of which row
corresponds to which outcome.}
}
\value{
List with two elements containing predictions for mu and sigma model
  terms. If a some model parameters are missing (such as sigma for binomial
  distributional assumption) NA elements are contained.
}
\description{
This is an internal function combining all mu and sigma outcomes,
respectively, taking into account the outcome information.
}
