% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_family.R
\name{fam}
\alias{fam}
\title{First draft of new family}
\usage{
fam(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
A bamlss family object.
}
\description{
Fix three distributional assumptions and do not supply any derivatives.
}
