% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_family.R
\name{fam2}
\alias{fam2}
\title{Next draft of new family}
\usage{
fam2(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
A bamlss family object.
}
\description{
Fix three distributional assumptions but supply derivatives.
}
