% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_family.R
\name{famg}
\alias{famg}
\title{Draft of new family for gamlss2}
\usage{
famg(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
A gamlss2 family object.
}
\description{
Fix three distributional assumptions but supply derivatives.
}
