% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{incorporate_outcome}
\alias{incorporate_outcome}
\title{Incorporate outcome information into}
\usage{
incorporate_outcome(pred_list, mus, sigmas, outcome_ids, outcome_levels)
}
\arguments{
\item{pred_list}{List of predictions for each outcome.}

\item{mus}{Integer vector for numbering available mus. Can be NULL but
shouldn't.}

\item{sigmas}{Integer vector for numbering available sigmas. Can be NULL.}

\item{outcome_ids}{Numeric matrix resulting from model.matrix call containing
the info about the outcomes. Column names are hard coded.}

\item{outcome_levels}{Character string containing the outcome names.}
}
\value{
List but now with 0 elements where the rows are not corresponding to
  outcomes.
}
\description{
This is an internal function multiplying all outcome predictions with 0 if
the respective row is not part of the outcome.
}
