% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_families.R
\name{trafficfam}
\alias{trafficfam}
\title{Draft of family for traffic example}
\usage{
trafficfam(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
A bamlss family object.
}
\description{
Fix four distributional assumptions and supply derivatives. Use BCCGo for
speed data. Use negative binomial for count data.
}
