% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_families.R
\name{trafficfam2}
\alias{trafficfam2}
\title{Draft 2 of family for traffic example}
\usage{
trafficfam2(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
A bamlss family object.
}
\description{
Fix four distributional assumptions and supply derivatives. Use BCCGo for
speed data. Use zero-truncated negative binomial for count data (no second
derivatives available).
}
