% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_families.R
\name{trafficfam4}
\alias{trafficfam4}
\title{Draft of family for traffic example}
\usage{
trafficfam4(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
A bamlss family object.
}
\description{
Fix four distributional assumptions and supply derivatives. Use lognormal for
speed data. Use Poisson for count data.
}
